Refactor Token Address Lookup Logic
